# openapi.yaml
openapi: 3.0.3
info:
  title: Professional Pride Room Booking API
  version: 2.0.0
  description: |
    Comprehensive REST API for room booking management system with advanced features including:
    - Room availability validation and conflict prevention
    - Customer management with ID proof verification
    - Dynamic pricing based on room configurations
    - Excel-style booking grid with horizontal scrolling
    - Payment processing with screenshot uploads
    - Invoice generation and PDF export
    - Real-time dashboard analytics
    - Mobile-optimized responsive interfaces
    
    **Key Features:**
    - Prevents double bookings with real-time validation
    - Supports multiple ID proof uploads per customer
    - Dynamic pricing based on number of guests
    - Comprehensive booking history tracking
    - Professional invoice generation
    - Mobile-responsive design
servers:
  - url: https://your-service.onrender.com
    description: Production server
  - url: http://localhost:8082
    description: Local development server
paths:
...
  /api/v1/bookings:
    get:
      summary: List all bookings
      description: Retrieve all bookings with comprehensive details
      responses:
        '200': 
          description: List of bookings
          content: 
            application/json: 
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Booking'
    post:
      summary: Create new booking
      description: |
        Create a new booking with automatic room availability validation.
        Returns 409 Conflict if room is already booked for the selected dates.
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/BookingCreateRequest'
      responses:
        '201': 
          description: Booking created successfully
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/Booking'
        '400': 
          description: Bad request - Invalid data
        '409': 
          description: Conflict - Room already booked for selected dates
        '500': 
          description: Internal server error
  /api/v1/bookings/{id}:
    get:
      summary: Get booking by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: integer
            format: int64
          description: Booking ID
      responses:
        '200': 
          description: Booking details
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/Booking'
        '404': 
          description: Booking not found
    put:
      summary: Update booking
      description: |
        Update booking with room availability validation.
        Returns 409 Conflict if updated dates conflict with existing bookings.
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/BookingUpdateRequest'
      responses:
        '200': 
          description: Booking updated successfully
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/Booking'
        '404': 
          description: Booking not found
        '409': 
          description: Conflict - Room already booked for updated dates
    delete:
      summary: Delete booking
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: integer
            format: int64
      responses:
        '204': 
          description: Booking deleted successfully
        '404': 
          description: Booking not found
  /api/v1/bookings/{id}/checkin:
    patch:
      summary: Check-in customer
      description: Change booking status from CONFIRMED to CHECKEDIN
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: integer
            format: int64
      responses:
        '200': 
          description: Customer checked in successfully
        '404': 
          description: Booking not found
        '400': 
          description: Invalid booking status for check-in
  /api/v1/bookings/{id}/checkout:
    patch:
      summary: Check-out customer
      description: Change booking status from CHECKEDIN to CHECKEDOUT
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: integer
            format: int64
      responses:
        '200': 
          description: Customer checked out successfully
        '404': 
          description: Booking not found
        '400': 
          description: Invalid booking status for check-out
  /api/v1/bookings/customer/{phoneNumber}:
    get:
      summary: Get customer booking history
      description: Retrieve all bookings for a specific customer
      parameters:
        - in: path
          name: phoneNumber
          required: true
          schema: 
            type: string
          description: Customer phone number
      responses:
        '200': 
          description: Customer booking history
          content: 
            application/json: 
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Booking'
  /api/v1/bookings/room/{roomId}:
    get:
      summary: Get room bookings
      description: Retrieve all bookings for a specific room
      parameters:
        - in: path
          name: roomId
          required: true
          schema: 
            type: integer
            format: int64
      responses:
        '200': 
          description: Room bookings
          content: 
            application/json: 
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Booking'
  /api/v1/bookings/status/{status}:
    get:
      summary: Get bookings by status
      parameters:
        - in: path
          name: status
          required: true
          schema: 
            $ref: '#/components/schemas/BookingStatus'
      responses:
        '200': 
          description: Bookings with specified status
          content: 
            application/json: 
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Booking'
...
  /api/v1/customer/{phoneNumber}:
    get:
      summary: Get customer by phone number
      parameters:
        - in: path
          name: phoneNumber
          required: true
          schema: 
            type: string
      responses:
        '200': 
          description: Customer details
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/Customer'
        '404': 
          description: Customer not found
    put:
      summary: Update customer
      parameters:
        - in: path
          name: phoneNumber
          required: true
          schema: 
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/CustomerUpdateRequest'
      responses:
        '200': 
          description: Customer updated successfully
        '404': 
          description: Customer not found
    delete:
      summary: Delete customer
      parameters:
        - in: path
          name: phoneNumber
          required: true
          schema: 
            type: string
      responses:
        '200': 
          description: Customer deleted successfully
        '404': 
          description: Customer not found
...
  /api/v1/payments/booking/{bookingId}:
    get:
      summary: Get payments for booking
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: 
            type: integer
            format: int64
      responses:
        '200': 
          description: Booking payments
          content: 
            application/json: 
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Payment'
...
components:
  schemas:
    Booking:
      type: object
      properties:
        id: 
          type: integer
          format: int64
          description: Unique booking identifier
        customerPhoneNumber: 
          type: string
          description: Customer phone number
        roomId: 
          type: integer
          format: int64
          description: Room identifier
        numberOfPeople: 
          type: integer
          description: Number of guests
        checkInDate: 
          type: string
          format: date-time
          description: Check-in date and time
        checkOutDate: 
          type: string
          format: date-time
          description: Check-out date and time
        bookingDurationType: 
          $ref: '#/components/schemas/BookingDurationType'
        bookingStatus: 
          $ref: '#/components/schemas/BookingStatus'
        paymentStatus: 
          $ref: '#/components/schemas/PaymentStatus'
        dailyCost: 
          type: number
          format: double
          description: Daily rate cost
        monthlyCost: 
          type: number
          format: double
          description: Monthly rate cost
        earlyCheckinCost: 
          type: number
          format: double
          description: Early check-in additional cost
        lateCheckoutCost: 
          type: number
          format: double
          description: Late check-out additional cost
        totalAmount: 
          type: number
          format: double
          description: Total booking amount
        remarks: 
          type: string
          description: Additional remarks
        createdAt: 
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt: 
          type: string
          format: date-time
          description: Last update timestamp
    BookingCreateRequest:
      type: object
      required: [customerPhoneNumber, roomId, checkInDate, checkOutDate, numberOfPeople]
      properties:
        customerPhoneNumber: 
          type: string
        roomId: 
          type: integer
          format: int64
        numberOfPeople: 
          type: integer
        checkInDate: 
          type: string
          format: date-time
        checkOutDate: 
          type: string
          format: date-time
        bookingDurationType: 
          $ref: '#/components/schemas/BookingDurationType'
        dailyCost: 
          type: number
          format: double
        monthlyCost: 
          type: number
          format: double
        earlyCheckinCost: 
          type: number
          format: double
        lateCheckoutCost: 
          type: number
          format: double
        remarks: 
          type: string
    BookingUpdateRequest:
      type: object
      properties:
        checkInDate: 
          type: string
          format: date-time
        checkOutDate: 
          type: string
          format: date-time
        numberOfPeople: 
          type: integer
        bookingStatus: 
          $ref: '#/components/schemas/BookingStatus'
        paymentStatus: 
          $ref: '#/components/schemas/PaymentStatus'
        dailyCost: 
          type: number
          format: double
        monthlyCost: 
          type: number
          format: double
        earlyCheckinCost: 
          type: number
          format: double
        lateCheckoutCost: 
          type: number
          format: double
        remarks: 
          type: string
...
    Customer:
      type: object
      properties:
        phoneNumber: 
          type: string
          description: Primary phone number (used as ID)
        name: 
          type: string
          description: Customer full name
        email: 
          type: string
          format: email
          description: Customer email address
        password: 
          type: string
          description: Customer password (hashed)
        additionalPhoneNumber: 
          type: string
          description: Additional contact number
        photoIdProofUrl: 
          type: string
          description: Legacy single ID proof URL
        idProofUrls: 
          type: array
          items: 
            type: string
          description: Multiple ID proof URLs
        idProofSubmitted: 
          type: boolean
          description: Whether ID proof has been submitted
        remarks: 
          type: string
          description: Additional remarks
        createdAt: 
          type: string
          format: date-time
        updatedAt: 
          type: string
          format: date-time
...
    CustomerUpdateRequest:
      type: object
      properties:
        name: 
          type: string
        email: 
          type: string
          format: email
        additionalPhoneNumber: 
          type: string
        remarks: 
          type: string
...
    Payment:
      type: object
      properties:
        id: 
          type: integer
          format: int64
        bookingId: 
          type: integer
          format: int64
        amount: 
          type: number
          format: double
          description: Payment amount
        paymentMethod: 
          $ref: '#/components/schemas/PaymentMode'
        paymentStatus: 
          $ref: '#/components/schemas/PaymentStatus'
        paymentScreenshotUrl: 
          type: string
          description: Screenshot of payment proof
        transactionId: 
          type: string
          description: Transaction reference ID
        paymentDate: 
          type: string
          format: date-time
          description: Payment date and time
        createdAt: 
          type: string
          format: date-time
        updatedAt: 
          type: string
          format: date-time
...
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, CHECKEDIN, CHECKEDOUT, CANCELLED, NO_SHOW, COMPLETED]
      description: |
        PENDING - Booking request submitted
        CONFIRMED - Booking confirmed by staff
        CHECKEDIN - Customer has checked in
        CHECKEDOUT - Customer has checked out
        CANCELLED - Booking cancelled
        NO_SHOW - Customer did not show up
        COMPLETED - Booking completed successfully
    BookingDurationType:
      type: string
      enum: [DAILY, MONTHLY]
      description: |
        DAILY - Daily rate booking
        MONTHLY - Monthly rate booking
    PaymentStatus:
      type: string
      enum: [PENDING, PAID, FAILED]
      description: |
        PENDING - Payment pending
        PAID - Payment completed
        FAILED - Payment failed
    PaymentMode:
      type: string
      enum: [CASH, UPI, CARD, BANK_TRANSFER]
      description: |
        CASH - Cash payment
        UPI - UPI payment
        CARD - Card payment
        BANK_TRANSFER - Bank transfer
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
security:
  - BearerAuth: []